refactor(credential)!: move AuditSink trait + audit types from storage into nebula-credential Core (ADR-0092)
AuditSink, AuditEvent, AuditOperation, AuditResult now live in
crates/credential/src/audit.rs and are re-exported flat at the
nebula-credential crate root.

AuditLayer (the decorator) stays in nebula-storage and imports the
trait+types from nebula_credential. The storage re-export chain
(layer/mod.rs -> credential/mod.rs) re-exports AuditLayer from the
local module and the four moved types directly from nebula_credential,
preserving the nebula_storage::credential::{AuditSink, ...} public
paths for all existing consumers (engine, credential-runtime, api)
without any import changes on their side.

Cargo.lock is unchanged (no new deps: chrono and StoreError were
already in nebula-credential).

cargo tree -p nebula-credential -e normal | grep "nebula-storage\|nebula-engine"
→ (empty — invariant holds)
